+++ game/player.py
@@ -3,7 +3,8 @@ from game.tiles import Tile
 
 class Player:
 
-    def __init__(self, bag_tiles):
+    def __init__(self,player_index,bag_tiles=BagTiles()):
+        self.player_index = player_index
         self.tiles = bag_tiles.take(7)
         self.bag_tiles = bag_tiles
     
@@ -18,5 +19,8 @@ class Player:
             if letter not in player_tiles_counts or count > player_tiles_counts[letter]:
                 return False
         return True
+    
+    def __str__(self):
+        return f"{self.player_index}"
     
     #def take_tile_using_letters()